pyproject.toml
[project]
name = "fastapi-doctr"
version = "0.1.0"
description = "FastAPI OCR extraction using Doctr"
authors = [{ name = "Vikram Dev", email = "vikramdev261@gmail.com" }]
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "fastapi",
    "uvicorn",
    "python-doctr",
    "pydantic",
    "python-multipart",
    "torch",
    "torchvision"
]

[tool.uvicorn]
host = "0.0.0.0"
port = 10000
reload = true